change python bindings build system, add more object functions to python bindings, libc write

+++ src/9p_proto.c
@@ -1321,8 +1321,8 @@ int p9p_read(struct p9_handle *p9_handle, struct p9_fid *fid, uint64_t offset, u
  * @param [IN]    data:		msk_registered msk_data pointer here
  * @return number of bytes written if >= 0, -errno on error.
  */
-int p9pz_write(struct p9_handle *p9_handle, struct p9_fid *fid, uint64_t offset, msk_data_t *data) {
- 	int rc;
+ssize_t p9pz_write(struct p9_handle *p9_handle, struct p9_fid *fid, uint64_t offset, msk_data_t *data) {
+ 	ssize_t rc;
 	msk_data_t *header_data;
 	uint16_t tag;
 	uint8_t msgtype;
@@ -1397,8 +1397,8 @@ int p9pz_write(struct p9_handle *p9_handle, struct p9_fid *fid, uint64_t offset,
  * @return number of bytes write if >= 0, -errno on error.
  *          0 indicates eof?
  */
-int p9p_write(struct p9_handle *p9_handle, struct p9_fid *fid, uint64_t offset, uint32_t count, char *buf) {
- 	int rc;
+ssize_t p9p_write(struct p9_handle *p9_handle, struct p9_fid *fid, uint64_t offset, size_t count, char *buf) {
+ 	ssize_t rc;
 	msk_data_t *data;
 	uint16_t tag;
 	uint8_t msgtype;
